Standalone psp->falco rules converter in go
Break out the PSP to falco rules converter to a standalone program,
written in go. It takes a PSP as input and writes a falco rules file as
output. It uses golang templating to render the set of rules and uses
k8s PodSecurityPolicy objects to hold the PSP after being read.

The templating is fairly straightforward, enabling/disabling rules based
on values in the PSP. As a convienence, there are several "Join"
functions that join lists of volumes, id ranges, port ranges, etc.

There's also a function to handle "is AllowPrivilegeEscalation set". It
has to be a function because go deferences pointers in templates to come
up with values, but AllowPrivilegeEscalation is true if unset.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>